naver_commerce_proxy/requirements.txt
fastapi
uvicorn
requests
bcrypt
python-dotenv
